requirements.txt
fastapi
uvicorn
pillow
torch
diffusers
numpy
opencv-python
mediapipe